+++ estimagic/parameters/check_constraints.py
