+++ estimagic/parameters/consolidate_constraints.py
