+++ estimagic/parameters/kernel_transformations.py
