+++ estimagic/parameters/parameter_conversion.py
@@ -9,14 +9,14 @@ High level means:
 """
 import functools
 
-from estimagic.optimization.process_constraints import process_constraints
-from estimagic.optimization.reparametrize import convert_external_derivative_to_internal
-from estimagic.optimization.reparametrize import post_replace_jacobian
-from estimagic.optimization.reparametrize import pre_replace_jacobian
-from estimagic.optimization.reparametrize import reparametrize_from_internal
-from estimagic.optimization.reparametrize import reparametrize_to_internal
 from estimagic.parameters.parameter_preprocessing import add_default_bounds_to_params
 from estimagic.parameters.parameter_preprocessing import check_params_are_valid
+from estimagic.parameters.process_constraints import process_constraints
+from estimagic.parameters.reparametrize import convert_external_derivative_to_internal
+from estimagic.parameters.reparametrize import post_replace_jacobian
+from estimagic.parameters.reparametrize import pre_replace_jacobian
+from estimagic.parameters.reparametrize import reparametrize_from_internal
+from estimagic.parameters.reparametrize import reparametrize_to_internal
 
 
 def get_reparametrize_functions(
