+++ tox.ini
@@ -79,7 +79,7 @@ ignore =
     W605  ; ignores regex relevant escape sequences
 per-file-ignores =
     docs/source/conf.py:E501
-    estimagic/optimization/kernal_transformations.py:N806
+    estimagic/parameters/kernel_transformations.py:N806
 warn-symbols =
     pytest.mark.wip = Remove 'wip' mark for tests.
